warehouse-stock-management/etl_pipeline/requirements.txt
contourpy==1.3.3
cycler==0.12.1
et-xmlfile==2.0.0
fonttools==4.60.1
greenlet==3.2.4
iniconfig==2.3.0
jinja2==3.1.6
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.7
numpy==2.3.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pdfkit==1.0.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pygments==2.19.2
pylatex==1.4.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
six==1.17.0
sqlalchemy==2.0.44
typing-extensions==4.15.0
tzdata==2025.2
